Some worldgen and throne fixes
Added a new value for Naughty Throne worldsetting to allow automatic respawn of gifts even when used without killing Klaus. This is now the default when playing on the Winterlands preset exclusively, rest is unchanged.

Also fixed vanilla Loot Stash not despawning unless player wakes it.

Modified some room / tasks chances that were changed for internal tests but made it into an update idk when, oops.